+++ src/lib/types/tags.ts
@@ -1,0 +1,7 @@
+export type Tag = {
+  id: string;
+  user_id: string;
+  name: string;
+  color: string;
+  created_at: string;
+};
